apps/backend/pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "grantcomply-agent-api"
version = "0.1.0"
description = "GrantComply Agent API Service"
authors = ["GrantComply Team"]
package-mode = false
packages = [{include = "agent_api", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"

supabase = "^2.8.0"
httpx = "^0.25.2"
python-multipart = "^0.0.6"
python-dotenv = "^1.0.0"
structlog = "^23.2.0"
pymupdf = "^1.23.0"
openai = "^1.3.0"
langchain = "^0.1.0"
langchain-openai = "^0.0.5"
langchain-core = "^0.1.0"
[tool.poetry.group.dev.dependencies]
# Removed unused testing and linting dependencies
# Add back if needed: pytest, black, isort, flake8, mypy, pre-commit



# Removed tool configurations for unused dev dependencies
# Uncomment and configure if tools are added back:
# [tool.black], [tool.isort], [tool.mypy]
